eks-agentic-ai-platform/requirements.txt
# Pinned at 2025-04-28
langgraph==1.1.10
langchain_openai==1.2.1
langfuse==4.5.1
python-dotenv==1.2.2
fastapi==0.136.1
uvicorn==0.46.0
hypothesis==6.152.4
pytest==9.0.3
requests==2.33.1
pyyaml==6.0.3
